implemented monte carlo integrator

+++ src/image.cpp
@@ -47,12 +47,12 @@ std::vector<Eigen::Vector2d> image_t::sample_pixel(unsigned int _x, unsigned int
 {
 	//grid
 	// std::vector<Eigen::Vector2d> samples;
-	// float start_x = float(_x)/width + 1.0/(2*samples_per_pixel*width);
-	// float start_y = float(_y)/height + 1.0/(2*samples_per_pixel*height);
+	// double start_x = double(_x)/width + 1.0/(2*samples_per_pixel*width);
+	// double start_y = double(_y)/height + 1.0/(2*samples_per_pixel*height);
 	// for(int i=0;i<samples_per_pixel;i++){
 	// 	for(int j=0;j<samples_per_pixel;j++){
-	// 		float center_x = start_x + float(i)/(samples_per_pixel*width);
-	// 		float center_y = start_y + float(j)/(samples_per_pixel*height);
+	// 		double center_x = start_x + double(i)/(samples_per_pixel*width);
+	// 		double center_y = start_y + double(j)/(samples_per_pixel*height);
 	// 		samples.push_back(Eigen::Vector2d(center_x, center_y));
 	// 	}
 	// }
@@ -64,12 +64,12 @@ std::vector<Eigen::Vector2d> image_t::sample_pixel(unsigned int _x, unsigned int
 	std::uniform_real_distribution<double> distribution_y(-1.0/(2*samples_per_pixel*height),1.0/(2*samples_per_pixel*height));
 
 	std::vector<Eigen::Vector2d> samples;
-	float start_x = float(_x)/width + 1.0/(2*samples_per_pixel*width);
-	float start_y = float(_y)/height + 1.0/(2*samples_per_pixel*height);
+	double start_x = double(_x)/width + 1.0/(2*samples_per_pixel*width);
+	double start_y = double(_y)/height + 1.0/(2*samples_per_pixel*height);
 	for(int i=0;i<samples_per_pixel;i++){
 		for(int j=0;j<samples_per_pixel;j++){
-			float center_x = start_x + float(i)/(samples_per_pixel*width);
-			float center_y = start_y + float(j)/(samples_per_pixel*height);
+			double center_x = start_x + double(i)/(samples_per_pixel*width);
+			double center_y = start_y + double(j)/(samples_per_pixel*height);
 			samples.push_back(Eigen::Vector2d(center_x + distribution_x(generator), center_y + distribution_y(generator)));
 		}
 	}
